Break MessageBuilder file into multiple files
