+++ formatter/src/commonMain/kotlin/builder/CommitMessageBuilder.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2022 Gabriel Feo
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package builder
+
+internal interface CommitMessageBuilder {
+    fun appendSubject(subject: String)
+    fun appendBody(
+        body: String,
+        stripComments: Boolean,
+    )
+
+    fun build(): String
+}
